+++ packages/terra-form-select/src/terra-dev-site/doc/example/combobox/Invalid.jsx
@@ -6,7 +6,7 @@ import styles from '../FormSelectDocCommon.module.scss';
 const cx = classNames.bind(styles);
 
 const InvalidExample = () => (
-  <Combobox placeholder="Select a color" isInvalid className={cx('form-select')}>
+  <Combobox isInvalid className={cx('form-select')}>
     <Combobox.Option value="blue" display="Blue" />
     <Combobox.Option value="green" display="Green" />
     <Combobox.Option value="purple" display="Purple" />