+++ packages/terra-form-select/src/terra-dev-site/doc/example/multiple/MultiSelect.jsx
@@ -6,7 +6,7 @@ import styles from '../FormSelectDocCommon.module.scss';
 const cx = classNames.bind(styles);
 
 const MultiSelectExample = () => (
-  <MultiSelect placeholder="Select a color" className={cx('form-select')}>
+  <MultiSelect className={cx('form-select')}>
     <MultiSelect.Option value="blue" display="Blue" />
     <MultiSelect.Option value="green" display="Green" />
     <MultiSelect.Option value="purple" display="Purple" />