+++ packages/terra-form-select/src/terra-dev-site/doc/example/search/IsTouchAccessible.jsx
@@ -6,7 +6,7 @@ import styles from '../FormSelectDocCommon.module.scss';
 const cx = classNames.bind(styles);
 
 const IsTouchAccessibleExample = () => (
-  <SearchSelect placeholder="Select a color" className={cx('form-select')} isTouchAccessible>
+  <SearchSelect className={cx('form-select')} isTouchAccessible>
     <SearchSelect.Option value="blue" display="Blue" />
     <SearchSelect.Option value="green" display="Green" />
     <SearchSelect.Option value="purple" display="Purple" />