+++ packages/terra-form-select/src/terra-dev-site/doc/example/tag/Tag.jsx
@@ -6,7 +6,7 @@ import styles from '../FormSelectDocCommon.module.scss';
 const cx = classNames.bind(styles);
 
 const TagSelectExample = () => (
-  <TagSelect placeholder="Select a color" className={cx('form-select')}>
+  <TagSelect className={cx('form-select')}>
     <TagSelect.Option value="blue" display="Blue" />
     <TagSelect.Option value="green" display="Green" />
     <TagSelect.Option value="purple" display="Purple" />