+++ packages/terra-form-select/src/terra-dev-site/test/form-select/InvalidTag.test.jsx
@@ -7,7 +7,7 @@ const cx = classNames.bind(styles);
 
 const InvalidTag = () => (
   <div className={cx('content-wrapper')}>
-    <Select isInvalid placeholder="Select a color" variant="tag" id="tag">
+    <Select isInvalid variant="tag" id="tag">
       <Select.Option value="blue" display="Blue" />
       <Select.Option value="green" display="Green" />
       <Select.Option value="purple" display="Purple" />